kernel: update bcma and ssb to wireless-testing/master master-2013-07-03
This updates 025-bcma_backport.patch and 020-ssb_update.patch to a
version of ssb and bcma used in recent wireless-testing. I removed all
patches adding serial and nand flash support before doing so to not
conflict with the existing flash support in the brcm47xx target. This
update was done to easily update the wireless drivers afterwards.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>

SVN-Revision: 37261