CC: linux-firmware: fix rt73-usb-firmware package
Backport of r49037

Some parts of this package were named rt73-pci-firmware before which
looks like a copy and past error. This renames all parts to rt73-usb-
firmware and then the firmware gets build and the dependencies from the
package with the kernel module are also working correctly.

This fixes #22069

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>

SVN-Revision: 49038