CC: kernel: Support for new Archer C7 with gd25q128 chip
Archer C7 V2.0 units from December 2015 onwards ( serials starting 215C
) have changed flash chips to the gd25q128 chip, this is supported in
trunk but not presently in 15.05. I would like stable support for this
version so I've back ported the required fix from trunk and removed a
conflicting patch from bcm53xx I've tested the patch it and I'm able to
install a build of 15.05.1 on a new Archer C7 device with this patch.

Signed-off-by: John Marrett <johnf@zioncluster.ca>
Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>

SVN-Revision: 49220